+++ inst/shiny/plan2litchi/fun1.R
@@ -1,0 +1,6 @@
+makeAP(projectDir = input$projectDir,
+       surveyArea=file1$datapath,
+       useMP = TRUE,
+       demFn = file2$datapath,
+       maxFlightTime = input$maxFlightTime,
+       uavType = "dji_csv")