+++ NEWS.md
@@ -2,7 +2,7 @@
 new features:
   * change DJI flag from `djip3` to `dji_csv`
   * re-add DJI support
-  * add converter for Missionplaner/Qgroundcontrol survey flightplans to lichi csv
+  * add converter and shiny GUI for Missionplaner/Qgroundcontrol survey flightplans to lichi csv
 bugfixes:
   * fix several waypoint issues especially the zig zag behaviour starting and landing